+++ R/create_format.R
@@ -99,6 +99,16 @@
 #'     "Thuringia"                     = 16,
 #'     "East"                          = 11:16)
 #'
+#' # With interval formats you can also use the keywords "low" and "high" to
+#' # catch everything from the lowest to the highest values, in case one doesn't
+#' # know exactly what the lowest and highest values are.
+#' income. <- interval_format(
+#'     "Total"              = c("low", "high"),
+#'     "below 500"          = c("low", 499),
+#'     "500 to under 1000"  = 500:999,
+#'     "1000 to under 2000" = 1000:1999,
+#'     "2000 and more"      = c(2000, "high"))
+#'
 #' @rdname formats
 #'
 #' @keywords internal
@@ -147,6 +157,35 @@ interval_format <- function(...){
     # Get from - to value as vectors
     from <- sapply(ranges, function(x) min(x, na.rm = TRUE))
     to   <- sapply(ranges, function(x) max(x, na.rm = TRUE))
+
+    # Insert pseudo low and high numbers for keywords
+    if (is.character(to)){
+        # First check if there are any other words than low and high in the format. If yes, abort.
+        if (!any(c("low", "high") %in% to)){
+            message(" X ERROR: Unknown keyword found. Creating interval format will be aborted.")
+            return(NULL)
+        }
+
+        # Low always ends up in "to", because it is a character value and comes alphabetically after high
+        if ("low" %in% to){
+            # Swap the real not "low" value to "to" and insert a pseudo low number in "from"
+            to[to == "low"]  <- from[to == "low"]
+            from[from == to] <- -1e20
+
+            from <- as.numeric(from)
+        }
+
+        # If there is the "high" keyword, at this point it will always be in "to" either the same way as low,
+        # just because it is character or because it is swapped here by the low code above.
+        if ("high" %in% to){
+            # Enter a pseude high number for "to"
+            to[to == "high"] <- 1e20
+
+            from <- as.numeric(from)
+        }
+
+        to <- as.numeric(to)
+    }
 
     end_time <- round(difftime(Sys.time(), start_time, units = "secs"), 3)
     message("- - - 'interval_format' execution time: ", end_time, " seconds")